Remove unused auto/config file
This would be the right approach in theory, but the current build
scripts are so brittle that this does not work. They should be rewritten
to be usable without the very specific setup currently required.

This file was not take into account for standard lernstick builds only
because it was not copied to the build environment by the build_tmpfs.sh
script. Otherwise builds would have failed all along...